requirements.txt
pathlib==1.0.1
chardet==5.2.0
PyPDF2==3.0.1
python-docx==1.2.0
pdfplumber==0.11.8
PyAML==6.0.3
gensim==4.4.0
tqdm==4.66.1